refactor(changelog-github): move template helpers to render-template.ts
